Refactor Code Smells from SonarQube
