+++ components/buildings/floorPlans/H/SVGHall8.js
