+++ components/buildings/floorPlans/H/SVGHall9.js
