+++ components/buildings/floorPlans/generateFloorPlan.js
