+++ components/buildings/floorPlans/index.js
@@ -1,5 +1,5 @@
 /* istanbul ignore file */
-export { default as Hall8 } from './H/Hall8';
-export { default as Hall9 } from './H/Hall9';
-export { default as Vl1 } from './VL/Vl1';
-export { default as Vl2 } from './VL/Vl2';
+export { default as Hall8 } from './H/SVGHall8';
+export { default as Hall9 } from './H/SVGHall9';
+export { default as Vl1 } from './VL/SVGVl1';
+export { default as Vl2 } from './VL/SVGVl2';
