+++ components/mapSearchBar/index.js
@@ -10,7 +10,7 @@ import {
 import { SearchBar, Tooltip } from 'react-native-elements';
 import i18n from 'i18n-js';
 import styles from './styles';
-import SetLocaleContext from '../../localization-context';
+import SetLocaleContext from '../../SetLocaleContext';
 import burger from '../../assets/icons/burger.png';
 
 export default class searchBar extends Component {
