+++ components/menu/calendar/index.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
 import { createAppContainer, createSwitchNavigator } from 'react-navigation';
 // eslint-disable-next-line no-unused-vars
-import firebase from '../../../firebase.config';
 import LoginScreen from './Screens/LoginScreen';
 import LoadingScreen from './Screens/LoadingScreen';
 import DashboardScreen from './Screens/DashboardScreen';
