+++ indoor_directions_modules/generateGraph.js
