+++ tests/indoorShortestPath.test.js
@@ -6,7 +6,7 @@ import BuildingWithFloors from '../components/buildings/buildingView/buildingWit
 import dijkstraPathfinder from '../indoor_directions_modules/dijkstraPathfinder';
 
 beforeEach(() => {
-  mockGraphFloor1 = {
+  const mockGraphFloor1 = {
     101: {
       x: 1,
       y: 1,
@@ -82,7 +82,7 @@ beforeEach(() => {
       ]
     }
   };
-  mockGraphFloor2 = {
+  const mockGraphFloor2 = {
     201: {
       x: 2,
       y: 1,
