+++ tests/theMap.test.js
@@ -42,7 +42,7 @@ it('Should fit screen to updated components', () => {
   };
 
   const fitToCoordinates = jest.fn();
-  wrapper = shallow(<TheMap nearbyMarkers={[]} updatedCoordinates={coordinatesPre} updatedRegion={mockRegion} />);
+  const wrapper = shallow(<TheMap nearbyMarkers={[]} updatedCoordinates={coordinatesPre} updatedRegion={mockRegion} />);
   const spyFitToCoordinates = jest.spyOn(wrapper.instance(), 'fitScreenToPath');
   wrapper.instance().setState({ mapRef: { fitToCoordinates } });
   wrapper.setProps({ updatedCoordinates: coordinatesChanged });
@@ -50,7 +50,7 @@ it('Should fit screen to updated components', () => {
 });
 
 it('Should fetch and send the selected point of interest to the home component', async () => {
-  component = shallow(<TheMap nearbyMarkers={[]} updatedRegion={mockRegion} getDestinationIfSet={getDestinationIfSet} updateRegionCloser={updateRegionCloser} />);
+  const component = shallow(<TheMap nearbyMarkers={[]} updatedRegion={mockRegion} getDestinationIfSet={getDestinationIfSet} updateRegionCloser={updateRegionCloser} />);
   const spyGetDestinationIfSet = jest.spyOn(component.instance().props, 'getDestinationIfSet');
   const spyUpdateRegionCloser = jest.spyOn(component.instance().props, 'updateRegionCloser');
   global.fetch = jest.fn().mockImplementation(() => {
